Moving BillingMessages to lib constants
