+++ src/api/modules/utils.ts
@@ -8,6 +8,7 @@ import { existsSync } from 'fs-extra'
 import { resolve as resolvePath } from 'path'
 import R, { compose, concat, contains, curry, head, last, prop, propSatisfies, reduce, split, tail, __ } from 'ramda'
 import semverDiff from 'semver-diff'
+import { BillingMessages } from '../../lib/constants/BillingMessages'
 import { createAppsClient } from '../clients/IOClients/infra/Apps'
 import { createRegistryClient } from '../clients/IOClients/infra/Registry'
 import { createWorkspacesClient } from '../clients/IOClients/infra/Workspaces'
@@ -17,7 +18,6 @@ import { ManifestEditor } from '../manifest'
 import { getAppRoot } from '../manifest/ManifestUtil'
 import { SessionManager } from '../session/SessionManager'
 import { createTable } from '../table'
-import { BillingMessages } from './billingMessages'
 import { promptConfirm } from './prompts'
 
 const workspaceExampleName = process.env.USER || 'example'
