+++ src/lib/constants/BillingMessages.ts
@@ -1,5 +1,5 @@
 import chalk from 'chalk'
-import { COLORS } from '..'
+import { COLORS } from '../../api'
 
 export const BillingMessages = {
   APP_STORE_TERMS_OF_SERVICE: chalk.hex(COLORS.YELLOW).bold('App Terms of Service: '),
@@ -8,6 +8,8 @@ export const BillingMessages = {
   SUBSCRIPTION_MONTHLY: 'Subscription (monthly)',
   BILLABLE_ITEM: chalk.bold('Billable item'),
   PRICING: chalk.bold('Pricing'),
+  INSTALL_STARTED: 'Starting to install app with accepted Terms',
+  INSTALL_SUCCESS: 'Installed after accepted terms',
   licenseLink: (url: string) => chalk.underline(url),
   price: (amount: number, currency: string) => `${amount} (${currency})`,
   pricePerUnit: (amount: number, currency = 'BRL') => `${BillingMessages.price(amount, currency)} per unit`,
@@ -16,6 +18,4 @@ export const BillingMessages = {
     `${chalk.hex(COLORS.BLUE)(app)} is a ${
       free ? chalk.hex(COLORS.GREEN)('free') : chalk.hex(COLORS.MAGENTA)('paid')
     } app. To install it, you need to accept the following Terms:\n\n${terms}\n`,
-  INSTALL_STARTED: 'Starting to install app with accepted Terms',
-  INSTALL_SUCCESS: 'Installed after accepted terms',
 }
